LPS-179353 Extract ScreenNavigationEntry methods from CommerceTaxMethodDetailScreenNavigationCategory to CommerceTaxMethodDetailScreenNavigationEntry, extends CommerceTaxMethodDetailScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
