+++ modules/apps/commerce/commerce-tax-web/src/main/java/com/liferay/commerce/tax/web/internal/frontend/taglib/servlet/taglib/CommerceTaxMethodDetailScreenNavigationCategory.java
@@ -15,19 +15,10 @@
 package com.liferay.commerce.tax.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.commerce.constants.CommerceTaxScreenNavigationConstants;
-import com.liferay.commerce.tax.model.CommerceTaxMethod;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
 
-import java.io.IOException;
-
 import java.util.Locale;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -36,15 +27,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceTaxMethodDetailScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceTaxMethod> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -53,14 +40,8 @@ public class CommerceTaxMethodDetailScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return CommerceTaxScreenNavigationConstants.
-			ENTRY_KEY_COMMERCE_TAX_METHOD_DETAIL;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(
+		return language.get(
 			locale,
 			CommerceTaxScreenNavigationConstants.
 				ENTRY_KEY_COMMERCE_TAX_METHOD_DETAIL);
@@ -72,24 +53,7 @@ public class CommerceTaxMethodDetailScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_TAX_METHOD;
 	}
 
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/commerce_tax_method/detail.jsp");
-	}
-
 	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(target = "(osgi.web.symbolicname=com.liferay.commerce.tax.web)")
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
