+++ modules/apps/commerce/commerce-tax-web/src/main/java/com/liferay/commerce/tax/web/internal/frontend/taglib/servlet/taglib/CommerceTaxMethodDetailScreenNavigationEntry.java
@@ -1,0 +1,63 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.tax.web.internal.frontend.taglib.servlet.taglib;
+
+import com.liferay.commerce.tax.model.CommerceTaxMethod;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
+import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
+
+import java.io.IOException;
+
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(
+	property = "screen.navigation.entry.order:Integer=10",
+	service = ScreenNavigationEntry.class
+)
+public class CommerceTaxMethodDetailScreenNavigationEntry
+	extends CommerceTaxMethodDetailScreenNavigationCategory
+	implements ScreenNavigationEntry<CommerceTaxMethod> {
+
+	@Override
+	public String getEntryKey() {
+		return getCategoryKey();
+	}
+
+	@Override
+	public void render(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws IOException {
+
+		_jspRenderer.renderJSP(
+			_servletContext, httpServletRequest, httpServletResponse,
+			"/commerce_tax_method/detail.jsp");
+	}
+
+	@Reference
+	private JSPRenderer _jspRenderer;
+
+	@Reference(target = "(osgi.web.symbolicname=com.liferay.commerce.tax.web)")
+	private ServletContext _servletContext;
+
+}
